Better optionality.
- More Y/N in Optional, making it easier to omit parameters.
- AT removed from Envelope, since T already gives it.

+++ src/commands/GetInfoCommand.cpp
@@ -310,7 +310,7 @@ bool GetInfoCommand::SendEnvelopes(const CommandContext &context)
             context.StartField( "points" );
             context.StartArray();
             double offset = pEnv->mOffset;
-            for( int k=0;k<pEnv->mEnv.size(); k++)
+            for( size_t k=0;k<pEnv->mEnv.size(); k++)
             {
                context.StartStruct( );
                context.AddItem( pEnv->mEnv[k].GetT()+offset, "t" );
